Cleaned up tests directories.
